+++ packages/core/CHANGELOG.md
@@ -10,9 +10,13 @@ The following is a curated list of changes in the Enact core module, newest chan
 
 ### Changed
 
+- `@core/handle.forwardWithPrevent` to also call `preventDefault()` on the originating event
+
 ### Fixed
 
-### Removed
+## [1.4.1] - 2017-07-05
+
+No significant changes.
 
 ## [1.4.0] - 2017-06-29